+++ yabapi-core/src/commonMain/kotlin/moe/sdl/yabapi/data/article/ArticleMedia.kt
@@ -6,12 +6,12 @@ import kotlinx.serialization.Serializable
 @Serializable
 public data class ArticleMedia(
     @SerialName("score") val score: Int? = null,
-    @SerialName("media_id") val mediaId: Int? = null,
+    @SerialName("media_id") val mediaId: Long? = null,
     @SerialName("title") val title: String? = null,
     @SerialName("cover") val cover: String? = null,
     @SerialName("area") val area: String? = null,
-    @SerialName("type_id") val typeId: Int? = null,
+    @SerialName("type_id") val typeId: Long? = null,
     @SerialName("type_name") val typeName: String? = null,
     @SerialName("spoiler") val spoiler: Int? = null,
-    @SerialName("season_id") val seasonId: Int? = null,
+    @SerialName("season_id") val seasonId: Long? = null,
 )
